# backend/internal/repository/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.24.0

package sqlc

import (
	"database/sql"
	"time"
)

type Schedule struct {
	ID               int32
	ClientName       string
	StartAt          time.Time
	EndAt            time.Time
	Location         string
	Latitude         sql.NullFloat64
	Longitude        sql.NullFloat64
	ClockInAt        sql.NullTime
	ClockInLat       sql.NullFloat64
	ClockInLng       sql.NullFloat64
	ClockOutAt       sql.NullTime
	ClockOutLat      sql.NullFloat64
	ClockOutLng      sql.NullFloat64
	CreatedAt        time.Time
	UpdatedAt        time.Time
	ClockInVerified  sql.NullBool
	ClockOutVerified sql.NullBool
}

type Task struct {
	ID         int32
	ScheduleID int32
	Title      string
	Status     string
	Reason     sql.NullString
	CreatedAt  time.Time
	UpdatedAt  time.Time
}